Revert "Created more of front end stuff (HTML) and trying to set up..."
