--- package.json
@@ -1,28 +1,0 @@
-{
-  "name": "makersbnb",
-  "version": "1.0.0",
-  "description": "## MVP User Stories",
-  "main": "app.js",
-  "directories": {
-    "lib": "lib"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Raefey/MakersBNB.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/Raefey/MakersBNB/issues"
-  },
-  "homepage": "https://github.com/Raefey/MakersBNB#readme",
-  "dependencies": {
-    "express": "^4.16.3",
-    "mysql": "^2.15.0",
-    "pg-promise": "^8.2.2",
-    "requirejs": "^2.3.5"
-  }
-}
