+++ src/connection.js
@@ -1,0 +1,5 @@
+var pg = require(‘pg’);
+var connectionString = "postgres://userName:password@serverName/ip:port/nameOfDatabase";
+var pgClient = new pg.Client(connectionString);
+pgClient.connect();
+var query = pgClient.query("SELECT id from Customer where name = 'customername'");
