--- views/breaker.html
@@ -1,42 +1,0 @@
-<html>
-  <head>
-    <script src="src/breaker.js"></script>
-    <script src="src/maker.js"></script>
-    <script src="https://code.jquery.com/jquery-2.1.4.min.js"></script>
-    <script src="src/interface.js"></script>
-  </head>
-
-  <body>
-    <div id="advert">
-      <h1>Raefe's Crib</h1>
-      <h2>£999.00</h2>
-      <p>Rent my super expensive crib! That price is nightly, be-atch</p>
-    </div>
-
-    <div id="breaker-rent-form">
-      <form action="/confirmation" method="get">
-        <label>Name</label>
-        <br/>
-        <input type="text" id="breaker-name"></input>
-        <br/><br/>
-
-        <label>Email</label>
-        <br/>
-        <input type="text" id="breaker-email"></input>
-        <br/><br/>
-
-        <label>Message</label>
-        <br/>
-        <textarea type="text" id="breaker-message" rows="20" cols="50"></textarea>
-        <br/><br/>
-
-        <label>Dates</label>
-        <br/>
-        <input type="text" id="breaker-dates"></input>
-        <br/><br/>
-
-        <input type=submit value="RENT!">
-      </form>
-    </div>
-  </body>
-</html>
