--- views/confirmation.html
@@ -1,1 +1,0 @@
-<h1>Rental form submitted!</h1>
