+++ views/index.html
@@ -4,13 +4,20 @@
   </head>
 
   <body>
-    <form action="/maker">
-      <input type="submit" value="Maker"></input>
-    </form>
-
-    <form action="/breaker">
-      <input type="submit" value="Breaker"></input>
-    </form>
+    <section>
+      <script src="../src/breaker.js"></script>
+      <script src="../src/maker.js"></script>
+      <script src="https://code.jquery.com/jquery-2.1.4.min.js"></script>
+      <script src="../src/interface.js"></script>
+      <form>
+            <label for='room-info'>Room Info: </label>
+            <input id='room-info' name='room-info' type='text'></input>
+            <br>
+            <br>
+             <button class='button' type='submit' id="add-a-listing">Add listing</button>
+         </form>
+      <h1><span id='Listings'></span></h1>
+    </section>
 
   </body>
 </html>
